+++ packages/http/tsconfig.json
@@ -1,6 +1,7 @@
 {
   "extends": "../tsconfig.json",
   "compilerOptions": {
+    "resolveJsonModule": true,
     "paths": {
       "@stoplight/prism-core": ["./core/src"]
     }
